+++ CHANGELOG.md
@@ -6,9 +6,11 @@ Notable changes will be documented in this file. The project adheres to [Semanti
 Unreleased
 ----------
 
-* Updated Travis CI configuration
+* Changed hook from `contentPostProc-output` to `contentPostProc-all`
+* Support for TYPO3 10.4 and up via middleware
 * Dropped support for TYPO3 9.5 (breaking change)
 * Dropped support for PHP < 7.4 (breaking change)
+* Updated Travis CI configuration
 * Dropped support for AppVeyor
 
 1.0.0